requirements.txt
attrs==21.2.0
coverage==5.5
iniconfig==1.1.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
protobuf==3.17.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.17.1
six==1.16.0
tabulate==0.8.9
toml==0.10.2
wcwidth==0.2.5
